flask-backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-jsonschema-validator
flask-login
flask-admin
flask-mail
requests

numpy
spacy
#!python -m spacy download en_core_web_sm
flair
scipy